Release: Merge release into master from: release/2.6.0

+++ dojo/tool_product/views.py
@@ -1,13 +1,13 @@
 # #  product
 import logging
 from django.contrib import messages
+from django.core.exceptions import BadRequest
 from django.urls import reverse
 from django.shortcuts import render, get_object_or_404
-from django.utils import timezone
-from dojo.forms import DeleteToolProductSettingsForm, ToolProductSettingsForm, NoteForm
+from dojo.forms import DeleteToolProductSettingsForm, ToolProductSettingsForm
 from dojo.models import Product, Tool_Product_Settings
 from django.http import HttpResponseRedirect
-from dojo.utils import add_breadcrumb, Product_Tab
+from dojo.utils import Product_Tab
 from dojo.authorization.authorization_decorators import user_is_authorized
 from dojo.authorization.roles_permissions import Permissions
 
@@ -55,46 +55,11 @@ def all_tool_product(request, pid):
 
 
 @user_is_authorized(Product, Permissions.Product_Edit, 'pid')
-def view_tool_product(request, pid, ttid):
-    tool = Tool_Product_Settings.objects.get(pk=ttid)
-    notes = tool.notes.all()
-
-    if request.method == 'POST':
-        form = NoteForm(request.POST)
-        if form.is_valid():
-            new_note = form.save(commit=False)
-            new_note.author = request.user
-            new_note.date = timezone.now()
-            new_note.save()
-            tool.notes.add(new_note)
-            form = NoteForm()
-            # url = request.build_absolute_uri(reverse("view_test", args=(test.id,)))
-            # title="Test: %s on %s" % (test.test_type.name, test.engagement.product.name)
-            # process_notifications(request, new_note, url, title)
-            messages.add_message(
-                request,
-                messages.SUCCESS,
-                'Note added successfully.',
-                extra_tags='alert-success')
-    else:
-        form = NoteForm()
-
-    add_breadcrumb(
-        title="View Product Tool Configuration",
-        top_level=False,
-        request=request)
-
-    return render(request, 'dojo/view_tool_product.html', {
-        'tool': tool,
-        'notes': notes,
-        'form': form
-    })
-
-
-@user_is_authorized(Product, Permissions.Product_Edit, 'pid')
 def edit_tool_product(request, pid, ttid):
-    prod = get_object_or_404(Product, id=pid)
+    product = get_object_or_404(Product, id=pid)
     tool_product = Tool_Product_Settings.objects.get(pk=ttid)
+    if tool_product.product != product:
+        raise BadRequest(f'Product {pid} does not fit to product of Tool_Product {tool_product.product.id}')
 
     if request.method == 'POST':
         tform = ToolProductSettingsForm(request.POST, instance=tool_product)
@@ -119,7 +84,10 @@ def edit_tool_product(request, pid, ttid):
 @user_is_authorized(Product, Permissions.Product_Edit, 'pid')
 def delete_tool_product(request, pid, ttid):
     tool_product = Tool_Product_Settings.objects.get(pk=ttid)
-    prod = get_object_or_404(Product, id=pid)
+    product = get_object_or_404(Product, id=pid)
+    if tool_product.product != product:
+        raise BadRequest(f'Product {pid} does not fit to product of Tool_Product {tool_product.product.id}')
+
     if request.method == 'POST':
         tform = DeleteToolProductSettingsForm(
             request.POST, instance=tool_product)
